Completed Sql CODE & entities in java...

+++ OnlineShop/src/main/java/yesko/project/OnlineShop/dto/UserDto.java
@@ -3,13 +3,14 @@ package yesko.project.OnlineShop.dto;
 import lombok.*;
 import yesko.project.OnlineShop.entity.User;
 
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserDto {
-    private Integer id;
+    private Long id;
     private String username;
     private String firstName;
     private String lastName;
